adapt ogr part and ad version traits
