+++ src/gdal_2_2/spatial_ref/mod.rs
@@ -1,0 +1,2 @@
+#[cfg(test)]
+mod tests;
