+++ src/gdal_2_2/vector/mod.rs
@@ -1,0 +1,2 @@
+#[cfg(test)]
+mod tests;
