+++ src/gdal_3_0/spatial_ref/mod.rs
@@ -1,0 +1,5 @@
+mod srs;
+pub use srs::{SpatialRef_3_0};
+
+#[cfg(test)]
+mod tests;
