+++ src/gdal_3_0/vector/mod.rs
@@ -1,0 +1,2 @@
+#[cfg(test)]
+mod tests;
