+++ src/gdal_common/spatial_ref/srs.rs
@@ -67,12 +67,6 @@ impl CoordTransform {
         }
     }
 
-    #[deprecated(since = "0.3.1", note = "use `transform_coords` instead")]
-    pub fn transform_coord(&self, x: &mut [f64], y: &mut [f64], z: &mut [f64]) {
-        self.transform_coords(x, y, z)
-            .expect("Coordinate transform failed")
-    }
-
     pub fn to_c_hct(&self) -> OGRCoordinateTransformationH {
         self.inner
     }
@@ -102,7 +96,21 @@ impl PartialEq for SpatialRef {
 }
 
 impl SpatialRef {
-    pub fn new() -> Result<SpatialRef> {
+
+    pub fn clone_from_c_obj(c_obj: OGRSpatialReferenceH) -> Result<SpatialRef> {
+        let mut_c_obj = unsafe { gdal_sys::OSRClone(c_obj) };
+        if mut_c_obj.is_null() {
+            Err(_last_null_pointer_err("OSRClone"))?
+        } else {
+            Ok(SpatialRef(mut_c_obj))
+        }
+    }
+}
+
+pub trait SpatialRefCommon {
+    fn c_spatial_ref(&self) -> OGRSpatialReferenceH;
+    
+    fn new() -> Result<SpatialRef> {
         let c_obj = unsafe { gdal_sys::OSRNewSpatialReference(ptr::null()) };
         if c_obj.is_null() {
             Err(_last_null_pointer_err("OSRNewSpatialReference"))?;
@@ -110,7 +118,7 @@ impl SpatialRef {
         Ok(SpatialRef(c_obj))
     }
 
-    pub fn from_definition(definition: &str) -> Result<SpatialRef> {
+    fn from_definition(definition: &str) -> Result<SpatialRef> {
         let c_obj = unsafe { gdal_sys::OSRNewSpatialReference(ptr::null()) };
         if c_obj.is_null() {
             Err(_last_null_pointer_err("OSRNewSpatialReference"))?;
@@ -126,7 +134,7 @@ impl SpatialRef {
         Ok(SpatialRef(c_obj))
     }
 
-    pub fn from_wkt(wkt: &str) -> Result<SpatialRef> {
+    fn from_wkt(wkt: &str) -> Result<SpatialRef> {
         let c_str = CString::new(wkt)?;
         let c_obj = unsafe { gdal_sys::OSRNewSpatialReference(c_str.as_ptr()) };
         if c_obj.is_null() {
@@ -135,7 +143,7 @@ impl SpatialRef {
         Ok(SpatialRef(c_obj))
     }
 
-    pub fn from_epsg(epsg_code: u32) -> Result<SpatialRef> {
+    fn from_epsg(epsg_code: u32) -> Result<SpatialRef> {
         let null_ptr = ptr::null_mut();
         let c_obj = unsafe { gdal_sys::OSRNewSpatialReference(null_ptr) };
         let rv = unsafe { gdal_sys::OSRImportFromEPSG(c_obj, epsg_code as c_int) };
@@ -149,7 +157,7 @@ impl SpatialRef {
         }
     }
 
-    pub fn from_proj4(proj4_string: &str) -> Result<SpatialRef> {
+    fn from_proj4(proj4_string: &str) -> Result<SpatialRef> {
         let c_str = CString::new(proj4_string)?;
         let null_ptr = ptr::null_mut();
         let c_obj = unsafe { gdal_sys::OSRNewSpatialReference(null_ptr) };
@@ -164,7 +172,7 @@ impl SpatialRef {
         }
     }
 
-    pub fn from_esri(esri_wkt: &str) -> Result<SpatialRef> {
+    fn from_esri(esri_wkt: &str) -> Result<SpatialRef> {
         let c_str = CString::new(esri_wkt)?;
         let mut ptrs = vec![c_str.as_ptr() as *mut i8, ptr::null_mut()];
         let null_ptr = ptr::null_mut();
@@ -180,18 +188,9 @@ impl SpatialRef {
         }
     }
 
-    pub fn from_c_obj(c_obj: OGRSpatialReferenceH) -> Result<SpatialRef> {
-        let mut_c_obj = unsafe { gdal_sys::OSRClone(c_obj) };
-        if mut_c_obj.is_null() {
-            Err(_last_null_pointer_err("OSRClone"))?
-        } else {
-            Ok(SpatialRef(mut_c_obj))
-        }
-    }
-
-    pub fn to_wkt(&self) -> Result<String> {
+    fn to_wkt(&self) -> Result<String> {
         let mut c_wkt = ptr::null_mut();
-        let rv = unsafe { gdal_sys::OSRExportToWkt(self.0, &mut c_wkt) };
+        let rv = unsafe { gdal_sys::OSRExportToWkt(self.c_spatial_ref(), &mut c_wkt) };
         if rv != OGRErr::OGRERR_NONE {
             Err(ErrorKind::OgrError {
                 err: rv,
@@ -202,8 +201,8 @@ impl SpatialRef {
         }
     }
 
-    pub fn morph_to_esri(&self) -> Result<()> {
-        let rv = unsafe { gdal_sys::OSRMorphToESRI(self.0) };
+    fn morph_to_esri(&self) -> Result<()> {
+        let rv = unsafe { gdal_sys::OSRMorphToESRI(self.c_spatial_ref()) };
         if rv != OGRErr::OGRERR_NONE {
             Err(ErrorKind::OgrError {
                 err: rv,
@@ -213,9 +212,9 @@ impl SpatialRef {
         Ok(())
     }
 
-    pub fn to_pretty_wkt(&self) -> Result<String> {
+    fn to_pretty_wkt(&self) -> Result<String> {
         let mut c_wkt = ptr::null_mut();
-        let rv = unsafe { gdal_sys::OSRExportToPrettyWkt(self.0, &mut c_wkt, false as c_int) };
+        let rv = unsafe { gdal_sys::OSRExportToPrettyWkt(self.c_spatial_ref(), &mut c_wkt, false as c_int) };
         if rv != OGRErr::OGRERR_NONE {
             Err(ErrorKind::OgrError {
                 err: rv,
@@ -226,9 +225,9 @@ impl SpatialRef {
         }
     }
 
-    pub fn to_xml(&self) -> Result<String> {
+    fn to_xml(&self) -> Result<String> {
         let mut c_raw_xml = ptr::null_mut();
-        let rv = unsafe { gdal_sys::OSRExportToXML(self.0, &mut c_raw_xml, ptr::null()) };
+        let rv = unsafe { gdal_sys::OSRExportToXML(self.c_spatial_ref(), &mut c_raw_xml, ptr::null()) };
         if rv != OGRErr::OGRERR_NONE {
             Err(ErrorKind::OgrError {
                 err: rv,
@@ -239,9 +238,9 @@ impl SpatialRef {
         }
     }
 
-    pub fn to_proj4(&self) -> Result<String> {
+    fn to_proj4(&self) -> Result<String> {
         let mut c_proj4str = ptr::null_mut();
-        let rv = unsafe { gdal_sys::OSRExportToProj4(self.0, &mut c_proj4str) };
+        let rv = unsafe { gdal_sys::OSRExportToProj4(self.c_spatial_ref(), &mut c_proj4str) };
         if rv != OGRErr::OGRERR_NONE {
             Err(ErrorKind::OgrError {
                 err: rv,
@@ -253,8 +252,8 @@ impl SpatialRef {
         }
     }
 
-    pub fn auth_name(&self) -> Result<String> {
-        let c_ptr = unsafe { gdal_sys::OSRGetAuthorityName(self.0, ptr::null()) };
+    fn auth_name(&self) -> Result<String> {
+        let c_ptr = unsafe { gdal_sys::OSRGetAuthorityName(self.c_spatial_ref(), ptr::null()) };
         if c_ptr.is_null() {
             Err(_last_null_pointer_err("SRGetAuthorityName"))?
         } else {
@@ -262,8 +261,8 @@ impl SpatialRef {
         }
     }
 
-    pub fn auth_code(&self) -> Result<i32> {
-        let c_ptr = unsafe { gdal_sys::OSRGetAuthorityCode(self.0, ptr::null()) };
+    fn auth_code(&self) -> Result<i32> {
+        let c_ptr = unsafe { gdal_sys::OSRGetAuthorityCode(self.c_spatial_ref(), ptr::null()) };
         if c_ptr.is_null() {
             Err(_last_null_pointer_err("OSRGetAuthorityCode"))?;
         }
@@ -278,13 +277,13 @@ impl SpatialRef {
         }
     }
 
-    pub fn authority(&self) -> Result<String> {
-        let c_ptr = unsafe { gdal_sys::OSRGetAuthorityName(self.0, ptr::null()) };
+    fn authority(&self) -> Result<String> {
+        let c_ptr = unsafe { gdal_sys::OSRGetAuthorityName(self.c_spatial_ref(), ptr::null()) };
         if c_ptr.is_null() {
             Err(_last_null_pointer_err("SRGetAuthorityName"))?;
         }
         let name = unsafe { CStr::from_ptr(c_ptr) }.to_str()?;
-        let c_ptr = unsafe { gdal_sys::OSRGetAuthorityCode(self.0, ptr::null()) };
+        let c_ptr = unsafe { gdal_sys::OSRGetAuthorityCode(self.c_spatial_ref(), ptr::null()) };
         if c_ptr.is_null() {
             Err(_last_null_pointer_err("OSRGetAuthorityCode"))?;
         }
@@ -292,8 +291,8 @@ impl SpatialRef {
         Ok(format!("{}:{}", name, code))
     }
 
-    pub fn auto_identify_epsg(&mut self) -> Result<()> {
-        let rv = unsafe { gdal_sys::OSRAutoIdentifyEPSG(self.0) };
+    fn auto_identify_epsg(&mut self) -> Result<()> {
+        let rv = unsafe { gdal_sys::OSRAutoIdentifyEPSG(self.c_spatial_ref()) };
         if rv != OGRErr::OGRERR_NONE {
             Err(ErrorKind::OgrError {
                 err: rv,
@@ -303,21 +302,11 @@ impl SpatialRef {
             Ok(())
         }
     }
+}
 
-    #[cfg(feature = "gdal_3_0")]
-    pub fn set_axis_mapping_strategy(&self, strategy: gdal_sys::OSRAxisMappingStrategy::Type) {
-        unsafe {
-            gdal_sys::OSRSetAxisMappingStrategy(self.0, strategy);
-        }
-    }
-
-    #[cfg(feature = "gdal_3_0")]
-    pub fn get_axis_mapping_strategy(&self) -> gdal_sys::OSRAxisMappingStrategy::Type {
-        unsafe { gdal_sys::OSRGetAxisMappingStrategy(self.0) }
-    }
-
-    // TODO: should this take self instead of &self?
-    pub fn to_c_hsrs(&self) -> OGRSpatialReferenceH {
+impl SpatialRefCommon for SpatialRef {
+    fn c_spatial_ref(&self) -> OGRSpatialReferenceH {
         self.0
     }
 }
+
