+++ src/gdal_common/vector/layer.rs
@@ -1,9 +1,6 @@
 use crate::gdal_common::gdal_major_object::MajorObject;
-use crate::metadata::Metadata;
-use crate::spatial_ref::SpatialRef;
 use crate::utils::{_last_null_pointer_err, _string};
-use crate::vector::defn::Defn;
-use crate::vector::{Feature, FieldValue, Geometry};
+use crate::vector::{Feature, FieldValue, Geometry, Defn};
 use gdal_sys::{
     self, GDALMajorObjectH, OGREnvelope, OGRErr, OGRFieldDefnH, OGRFieldType, OGRLayerH,
 };
@@ -11,13 +8,13 @@ use libc::c_int;
 use std::ffi::CString;
 use std::ptr::null_mut;
 
-use crate::{dataset::Dataset, errors::*};
+use crate::{Dataset, errors::*, Metadata, SpatialRef };
 
 /// Layer in a vector dataset
 ///
 /// ```
 /// use std::path::Path;
-/// use gdal::vector::Dataset;
+/// use gdal::{Dataset, DatasetCommon, vector::{VectorDatasetCommon, VectorLayerCommon}};
 ///
 /// let mut dataset = Dataset::open(Path::new("fixtures/roads.geojson")).unwrap();
 /// let layer = dataset.layer(0).unwrap();
@@ -150,7 +147,7 @@ pub trait VectorLayerCommon<'a> {
         if c_obj.is_null() {
             Err(_last_null_pointer_err("OGR_L_GetSpatialRef"))?;
         }
-        SpatialRef::from_c_obj(c_obj)
+        SpatialRef::clone_from_c_obj(c_obj)
     }
 }
